+++ backend/internal/models/notification.go
@@ -3,8 +3,8 @@ package models
 import "time"
 
 type Notification struct {
-	ID        int       `json:"id"`
-	UserID    int       `json:"userId"`
+	ID        int32     `json:"id"`
+	UserID    int64     `json:"userId"`
 	Content   string    `json:"content"`
 	Read      bool      `json:"read"`
 	CreatedAt time.Time `json:"createdAt"`
